requirements.txt
agency-swarm
aiofiles==24.1.0
Crawl4AI==0.4.247
html2text==2.0.0
httpx==0.27.2
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
tiktoken==0.8.0
gradio